Conflicts:
	src/main/frontend/src/app/helpers/loader.interceptor.ts
	src/main/java/com/apmosys/employeeportal/service/DepartmentService.java
	src/main/java/com/apmosys/employeeportal/service/JobRoleService.java

+++ src/main/java/com/apmosys/employeeportal/service/JobRoleService.java
@@ -13,8 +13,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.interceptor.TransactionAspectSupport;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
+import org.springframework.web.client.HttpServerErrorException.InternalServerError;
 
+import com.apmosys.employeeportal.dto.DepartmentDTO;
 import com.apmosys.employeeportal.dto.JobRoleDTO;
 import com.apmosys.employeeportal.dto.LogDTO;
 import com.apmosys.employeeportal.dto.PoPortalDTO;
@@ -340,6 +344,7 @@ public class JobRoleService {
 		return response;
 	}
 
+	@Transactional
 	public ServiceResponse deleteJobRole(JobRoleDTO jobRoleDTO) {
 		ServiceResponse response = new ServiceResponse();
 		LogDTO apiLogInfo = new LogDTO();
@@ -348,13 +353,63 @@ public class JobRoleService {
 		apiLogInfo.setLogLevel("INFO");
 		StringBuilder logBuilder = new StringBuilder();
 		logBuilder.append("jobRoleId : " + jobRoleDTO.getJobRoleId());
+		
+		boolean isJobRoleUsedInPoPortal = false;
 		try {
 			Optional<JobRole> jobRoleObject = jobRoleRepository.findById(jobRoleDTO.getJobRoleId());
 			if (jobRoleObject.isPresent()) {
 				JobRole jobRoleToBeDeleted = jobRoleObject.get();
 
 				Long count = employeeRepository.countByJobRoleId(jobRoleToBeDeleted.getJobRoleId());
-				if (count == 0) {
+				
+				//check JobRole in PoPortal
+				try {
+					
+					final String syncUrl = "http://192.168.21.175:8080/PoPortal/ishine/isRoleUsed/{roleId}";
+					RestTemplate restTemplate = new RestTemplate();
+					String syncResponse = restTemplate.getForObject(syncUrl, String.class, jobRoleToBeDeleted.getJobRoleId());
+					
+					JSONObject json = new JSONObject(syncResponse);
+					
+					if(json.getInt("httpStatusCode") == 200) {
+						
+						//0=Not Present
+						//1=Present but not used
+						//2=Used
+						
+						if(json.get("message").equals("0")) {
+							isJobRoleUsedInPoPortal = false;
+						}
+						
+						if(json.get("message").equals("1")) {
+							//Delete jobRole from poPortal http://192.168.21.175:8080/ishine/deleteRole/{id}
+							
+							final String deleteUrl = "http://192.168.21.175:8080/PoPortal/ishine/deleteRole/{id}";
+							RestTemplate deleteRestTemplate = new RestTemplate();
+							restTemplate.delete(deleteUrl, jobRoleToBeDeleted.getJobRoleId());
+						}
+						
+						if(json.get("message").equals("2")) {
+							isJobRoleUsedInPoPortal = true;
+						}
+					}
+					
+				}catch(InternalServerError e) {
+					TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+					JSONObject json = new JSONObject(e.getResponseBodyAsString());
+					
+					response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+					response.setServiceResponse(json.get("message"));
+					
+				}catch(HttpClientErrorException e) {
+					TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+                    JSONObject json = new JSONObject(e.getResponseBodyAsString());
+					
+					response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+					response.setServiceResponse(json.get("message"));
+				}
+				
+				if (count == 0 && isJobRoleUsedInPoPortal == false) {
 
 					jobRoleRepository.deleteById(jobRoleToBeDeleted.getJobRoleId());
 					response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
@@ -364,10 +419,24 @@ public class JobRoleService {
 					apiLogInfo.setApiStatus(ServiceResponse.STATUS_SUCCESS);
 					
 				} else {
-					response.setServiceStatus(ServiceResponse.STATUS_FAIL);
-					response.setServiceResponse("Job role cannot be deleted as it is mapped to employee.");
 					
-					apiLogInfo.setApiResponse("Job role cannot be deleted as it is mapped to employee.");			
+					JobRoleDTO dtoObject = new JobRoleDTO();
+
+					if (count != 0) {
+						dtoObject.setIsJobRoleUsedInIshine("true");
+					} else {
+						dtoObject.setIsJobRoleUsedInIshine("false");
+					}
+					if (isJobRoleUsedInPoPortal == true) {
+						dtoObject.setIsJobRoleUsedInPoPortal("true");
+					} else {
+						dtoObject.setIsJobRoleUsedInPoPortal("false");
+					}
+					
+					response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+					response.setServiceResponse(dtoObject);
+					
+					apiLogInfo.setApiResponse(dtoObject + "Job role cannot be deleted as it is mapped to employee.");			
 					apiLogInfo.setApiStatus(ServiceResponse.STATUS_FAIL);
 					
 				}
@@ -395,6 +464,7 @@ public class JobRoleService {
 		return response;
 	}
 
+	@Transactional
 	public ServiceResponse changeEmployeeJobRoleMapping(JobRoleDTO jobRoleDTO) {
 		ServiceResponse response = new ServiceResponse();
 		LogDTO apiLogInfo = new LogDTO();
@@ -406,61 +476,38 @@ public class JobRoleService {
 		
 		try {
 			List<Employee> employeeJobRole = employeeRepository.findByJobRoleId(jobRoleDTO.getOldJobRoleId());
-
 			if (!employeeJobRole.isEmpty()) {
 				for (Employee newJobRole : employeeJobRole) {
 
-					newJobRole.setJobRoleId(jobRoleDTO.getJobRoleId());
-
-					Employee dbResponse = employeeRepository.save(newJobRole);
+					Employee dbResponse = null;
+					if(jobRoleDTO.getIsJobRoleUsedInIshine().equals("true")) {
+						newJobRole.setJobRoleId(jobRoleDTO.getJobRoleId());
+						dbResponse = employeeRepository.save(newJobRole);
+					}else {
+						dbResponse.setName("Not used in Ishine");
+					}
 
 					if (dbResponse != null) {
-						
-//						http://192.168.21.175:8080/ishine/updateRole/{id}
-//							Type: POST
-//							PathVariable : id of job role to be deleted
-//							RequestBody:  New role to be replaced in place of deleted one.
-//							{
-//							    "roleId":10,
-//							    "roleName":"testRole",
-//							    "deptId":5
-//							}
-						
-						//Sync deleted jobRole with PoPortal
-						JobRole jobRoleObject = jobRoleRepository.findByjobRoleId(jobRoleDTO.getJobRoleId());
-						
-						if(jobRoleObject != null) {
-							JobRoleDTO syncObject = new JobRoleDTO();
+						if(jobRoleDTO.getIsJobRoleUsedInPoPortal().equals("true")) {
 							
-							syncObject.setRoleId(jobRoleObject.getJobRoleId());
-							syncObject.setRoleName(jobRoleObject.getName());
-							syncObject.setDeptId(jobRoleObject.getDeptId());
+							ServiceResponse deleteSyncResposne = syncDeleteJobRoleWithPoPortal(jobRoleDTO);
 							
-							final String syncUrl = syncJobRoleApi;
-							RestTemplate restTemplate = new RestTemplate();
-							String syncResponse = restTemplate.postForObject(syncUrl, syncObject, String.class, jobRoleDTO.getOldJobRoleId());
-							
-							JSONObject json = new JSONObject(syncResponse);
-							
-							if(json.getInt("httpStatusCode") == 200) {
+							if(deleteSyncResposne.getServiceStatus().equals("Success")) {
 								response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
-								response.setServiceResponse("Job Role deleted & Synced with PoPortal");
-								
-								apiLogInfo.setApiResponse("Job Role deleted & Synced with PoPortal");			
-								apiLogInfo.setApiStatus(ServiceResponse.STATUS_SUCCESS);
+								response.setServiceResponse(deleteSyncResposne.getServiceResponse());
 							}else {
-								response.setServiceStatus(ServiceResponse.STATUS_FAIL);
-								response.setServiceResponse("Job Role deleted, but unable to Sync with PoPortal");
+								TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
 								
-								apiLogInfo.setApiResponse("Job Role deleted, but unable to Sync with PoPortal");			
-								apiLogInfo.setApiStatus(ServiceResponse.STATUS_FAIL);
+								response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+								response.setServiceResponse(deleteSyncResposne.getServiceResponse());
 							}
-						}else {
-							response.setServiceStatus(ServiceResponse.STATUS_FAIL);
-							response.setServiceResponse("Job Role deleted, but unable to Sync with PoPortal");
 							
-							apiLogInfo.setApiResponse("Job Role deleted, but unable to Sync with PoPortal");			
-							apiLogInfo.setApiStatus(ServiceResponse.STATUS_FAIL);
+						}else {
+							response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
+							response.setServiceResponse("Job Role deleted successfully.");
+							
+							apiLogInfo.setApiResponse("Job Role deleted successfully.");			
+							apiLogInfo.setApiStatus(ServiceResponse.STATUS_SUCCESS);
 						}
 					} else {
 						response.setServiceStatus(ServiceResponse.STATUS_FAIL);
@@ -469,6 +516,24 @@ public class JobRoleService {
 						apiLogInfo.setApiResponse("Employee Job role mapping Failed.");			
 						apiLogInfo.setApiStatus(ServiceResponse.STATUS_FAIL);
 					}
+				}
+			}else {
+				//Check if poPortal is using jobRole
+				if(jobRoleDTO.getIsJobRoleUsedInPoPortal().equals("true")) {
+					ServiceResponse deleteSyncResposne = syncDeleteJobRoleWithPoPortal(jobRoleDTO);
+					if(deleteSyncResposne.getServiceStatus().equals("Success")) {
+						
+						response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
+						response.setServiceResponse(deleteSyncResposne.getServiceResponse());
+					}else {
+						TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+						
+						response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+						response.setServiceResponse(deleteSyncResposne.getServiceResponse());
+					}
+				}else {
+					response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+					response.setServiceResponse("No Employee found.");
 				}
 			}
 
@@ -483,6 +548,58 @@ public class JobRoleService {
 		}
 		apiLogInfo.setApiRequest(logBuilder.toString());
 		logService.logMyInfo(httpRequest, apiLogInfo);
+		return response;
+	}
+	
+	public ServiceResponse syncDeleteJobRoleWithPoPortal(JobRoleDTO jobRoleDTO) {
+		ServiceResponse response = new ServiceResponse();
+		try {
+			
+			//Sync deleted jobRole with PoPortal
+			JobRole jobRoleObject = jobRoleRepository.findByjobRoleId(jobRoleDTO.getJobRoleId());
+			
+			if(jobRoleObject != null) {
+				JobRoleDTO syncObject = new JobRoleDTO();
+				
+				syncObject.setRoleId(jobRoleObject.getJobRoleId());
+				syncObject.setRoleName(jobRoleObject.getName());
+				syncObject.setDeptId(jobRoleObject.getDeptId());
+				
+				try {
+					
+					final String syncUrl = "http://192.168.21.175:8080/PoPortal/ishine/updateRole/{id}";
+					RestTemplate restTemplate = new RestTemplate();
+					String syncResponse = restTemplate.postForObject(syncUrl, syncObject, String.class, jobRoleDTO.getOldJobRoleId());	
+					
+					JSONObject json = new JSONObject(syncResponse);
+					
+					if(json.getInt("httpStatusCode") == 200) {
+						response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
+						response.setServiceResponse("Job Role deleted & Synced with PoPortal");
+					}
+					
+				}catch(InternalServerError e) {
+					JSONObject json = new JSONObject(e.getResponseBodyAsString());
+					
+					response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+					response.setServiceResponse(json.get("message"));
+					
+				}catch(HttpClientErrorException e) {
+                    JSONObject json = new JSONObject(e.getResponseBodyAsString());
+					
+					response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+					response.setServiceResponse(json.get("message"));
+				}
+			}else {
+				response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+				response.setServiceResponse("JobRole not found.");
+			}
+		}catch(Exception e) {
+			e.printStackTrace();
+			response.setServiceStatus(ServiceResponse.SOMETHING_WENT_WRONG);
+			response.setServiceResponse("Something Went Wrong.");
+			response.setServiceError(e.getMessage());
+		}
 		return response;
 	}
